Consolidate ExpiryProbModel class, separate model logics from app
ExpiryProbModel is now consolidated into a separate class, for better
future integration. The expiry probability model is currently not
integrated into the main BN model used for success rate prediction

RecipeAI models is now separated from the RecipeAIApp. The RecipeAIApp
file now only handles codes required to run the app, with some user
interactions. The AI models are handled separately. This should make
the code better to read.
Utility functions and evidence injections are to be implemented under
the RecipeAI class.

Query now has changed. No evidence or variable is expected from the
user. Query result now return an array of recipe with corresponding
success rate. After the utility function is implemented, this is
expected to return the recipe with highest utility only.

Evidence injection is to be done within the class, not from the user.
More can be discussed regarding this.